// server/src/enums/onboarding.enum.ts
export enum LevelEnum {
  ZERO = 'ZERO',
  N5 = 'N5',
  N4 = 'N4',
  N3 = 'N3',
  N2 = 'N2',
  N1 = 'N1',
}

export enum StudyPaceEnum {
  RELAX = 'RELAX',
  DETERMINED = 'DETERMINED',
  RIGOROUS = 'RIGOROUS',
}

export enum StudyPreferenceEnum {
  GRAMMAR = 'GRAMMAR',
  VOCABULARY = 'VOCABULARY',
  BOTH = 'BOTH',
}